refactor: move manifest parser and spec to service layer
Relocates two manifest-related packages from internal/manifest to
internal/service/manifest per ADR 004 (layered architecture). Both
packages contain service-shaped orchestration:

- parser hosts CachedManifestParser (the cache wrapper consumed by the
  render Service) and Parser (template-to-module orchestration consumed
  by mandatory-module installation).
- spec hosts Resolver (constructor-injected service per ADR 002), the
  Spec data type, and the consumer-defined KeyChainLookup / PathExtractor
  interfaces that the OCI repository receives via DI.

Consumer-defined interfaces stay at their consumers per ADR 001:
- CachedManifestParser stays in render/service.go.
- SpecResolver, ResourceRenderService etc. stay in the manifest
  controller.

internal/manifest/statecheck is left in place: it is a pure stateless
K8s-resource-introspection helper, not a service. Renaming its types
to *Service to fit under internal/service/ would contradict ADR 005
("let the package name provide context"). Its only public interface,
StateCheck, is already consumer-defined in internal/controller/manifest.
