+++ internal/service/manifest/spec/resolver.go
