+++ internal/service/manifest/spec/spec.go
