+++ tests/integration/controller/manifest/suite_test.go
@@ -47,8 +47,8 @@ import (
 	"github.com/kyma-project/lifecycle-manager/internal/manifest/keychainprovider"
 	"github.com/kyma-project/lifecycle-manager/internal/manifest/labelsremoval"
 	"github.com/kyma-project/lifecycle-manager/internal/manifest/manifestclient"
-	"github.com/kyma-project/lifecycle-manager/internal/manifest/parser"
-	"github.com/kyma-project/lifecycle-manager/internal/manifest/spec"
+	"github.com/kyma-project/lifecycle-manager/internal/service/manifest/parser"
+	"github.com/kyma-project/lifecycle-manager/internal/service/manifest/spec"
 	"github.com/kyma-project/lifecycle-manager/internal/manifest/statecheck"
 	"github.com/kyma-project/lifecycle-manager/internal/pkg/metrics"
 	kymarepo "github.com/kyma-project/lifecycle-manager/internal/repository/kyma"
