+++ unit-test-coverage-lifecycle-manager.yaml
@@ -16,9 +16,7 @@ packages:
   internal/manifest/labelsremoval: 100
   internal/manifest/manifestclient: 5
   internal/manifest/modulecr: 56.9
-  internal/manifest/parser: 17.6
   internal/manifest/skrresources: 30.1
-  internal/manifest/spec: 92.3
   internal/manifest/statecheck: 74.2
   internal/manifest/status: 73.8
   internal/maintenancewindows: 91.2
@@ -45,7 +43,9 @@ packages:
   internal/service/kyma/status/modules/generator/fromerror: 100
   internal/service/kyma/lookup: 100
   internal/service/manifest/orphan: 100
+  internal/service/manifest/parser: 17.6
   internal/service/manifest/render: 92.5
+  internal/service/manifest/spec: 92.3
   internal/service/watcher/certificate: 100
   internal/service/watcher/certificate/secret/data: 100
   internal/service/watcher/chartreader: 100
